FlaskMicroservice/student-enrollmentWithDbFIIGNYA/requirements.txt
Flask
requests
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
psycopg2-binary